new sparse implementation; new profiling code

+++ ateams/arithmetic/common.py
@@ -1,6 +1,6 @@
 
 import numpy as np
 
-SFINT = np.int8
-FINT = np.int8
+SFINT = np.int32
+FINT = np.int32
